+++ system_metrics/types.go
@@ -1,0 +1,65 @@
+package systemmetrics
+
+import "github.com/pkg/errors"
+
+// CompressionType describes how the system metrics data is compressed.
+type CompressionType int32
+
+// Valid CompressionType values.
+const (
+	CompressionTypeNone  CompressionType = 0
+	CompressionTypeTARGZ CompressionType = 1
+	CompressionTypeZIP   CompressionType = 2
+	CompressionTypeGZ    CompressionType = 3
+	CompressionTypeXZ    CompressionType = 4
+)
+
+func (f CompressionType) validate() error {
+	switch f {
+	case CompressionTypeNone, CompressionTypeTARGZ, CompressionTypeZIP, CompressionTypeGZ, CompressionTypeXZ:
+		return nil
+	default:
+		return errors.New("invalid compression type specified")
+	}
+}
+
+// SchemaType describes how the time series data is stored.
+type SchemaType int32
+
+// Valid SchemaType values.
+const (
+	SchemaTypeRawEvents             SchemaType = 0
+	SchemaTypeCollapsedEvents       SchemaType = 1
+	SchemaTypeIntervalSummarization SchemaType = 2
+	SchemaTypeHistogram             SchemaType = 3
+)
+
+func (f SchemaType) validate() error {
+	switch f {
+	case SchemaTypeRawEvents, SchemaTypeCollapsedEvents, SchemaTypeIntervalSummarization, SchemaTypeHistogram:
+		return nil
+	default:
+		return errors.New("invalid schema type specified")
+	}
+}
+
+// DataFormat describes how the time series data is stored.
+type DataFormat int32
+
+// Valid DataFormat values.
+const (
+	DataFormatText DataFormat = 0
+	DataFormatFTDC DataFormat = 1
+	DataFormatBSON DataFormat = 2
+	DataFormatJSON DataFormat = 3
+	DataFormatCSV  DataFormat = 4
+)
+
+func (f DataFormat) validate() error {
+	switch f {
+	case DataFormatText, DataFormatFTDC, DataFormatBSON, DataFormatJSON, DataFormatCSV:
+		return nil
+	default:
+		return errors.New("invalid schema type specified")
+	}
+}
